Move asset integration tests into "v2_1" namespace
