+++ tests/v2_1/Api/Asset/UpsertListAssetApiIntegration.php
@@ -1,61 +1,56 @@
 <?php
 
-namespace Akeneo\PimEnterprise\ApiClient\tests\Common\Api\AssetCategory;
+namespace Akeneo\PimEnterprise\ApiClient\tests\v2_1\Api\Asset;
 
 use Akeneo\PimEnterprise\ApiClient\tests\Common\Api\ApiTestCase;
 
-class UpsertListAssetCategoryApiIntegration extends ApiTestCase
+class UpsertListAssetApiIntegration extends ApiTestCase
 {
     public function testUpsertListSuccessful()
     {
-        $api = $this->createClient()->getAssetCategoryApi();
-        $response = $api->upsertList([
+        $api = $this->createClient()->getAssetApi();
+        $response  = $api->upsertList([
             [
-                'code' => 'asset_main_catalog',
-                'labels' => [
-                    'en_US' => 'Nullam ullamcorper',
-                ],
+                'code' => 'akeneo_logo',
+                'description' => 'Akeneo logo updated',
             ],
             [
-                'code' => 'asset_cold',
-                'parent' => 'asset_winter',
-                'labels' => [
-                    'en_US' => 'Aenean ultricies elit',
-                ],
+                'code' => 'unicorn',
+                'description' => 'Unicorn asset',
+                'localizable' => false,
+                'end_of_use' => null
             ]
         ]);
 
         $this->assertInstanceOf('\Iterator', $response);
-
         $responseLines = iterator_to_array($response);
         $this->assertCount(2, $responseLines);
 
         $this->assertSame([
             'line'        => 1,
-            'code'        => 'asset_main_catalog',
+            'code'        => 'akeneo_logo',
             'status_code' => 204,
         ], $responseLines[1]);
 
         $this->assertSame([
             'line'        => 2,
-            'code'        => 'asset_cold',
+            'code'        => 'unicorn',
             'status_code' => 201,
         ], $responseLines[2]);
     }
 
     public function testUpsertListFailed()
     {
-        $api = $this->createClient()->getChannelApi();
-        $response = $api->upsertList([
+        $api = $this->createClient()->getAssetApi();
+
+        $response  = $api->upsertList([
             [
-                'labels' => [
-                    'en_US' => 'Nullam ullamcorper',
-                ],
+                'description' => 'Upsert without code',
+                'categories' => [],
             ]
         ]);
 
         $this->assertInstanceOf('\Iterator', $response);
-
         $responseLines = iterator_to_array($response);
         $this->assertCount(1, $responseLines);
